fix save/load for night (float and int are ennemies)

+++ Assets/Scripts/SaveLoadSystem/SaveManager.cs
@@ -9,9 +9,9 @@ public class SaveManager : MonoBehaviour
         PlayerPrefs.Save();
     }
 
-    public void SaveNightNumber(float nightNumber)
+    public void SaveNightNumber(int nightNumber)
     {
-        PlayerPrefs.SetFloat("NightNumber", nightNumber);
+        PlayerPrefs.SetInt("NightNumber", nightNumber);
         PlayerPrefs.Save();
     }
